refactor(brett): ws-client.ts → ws-connection-client + ws-undo-state (T000661)
