+++ brett/src/client/ws-client.ts
@@ -9,6 +9,10 @@ import { createLobbyState, applyLobbyServerMessage, type LobbyState } from './lo
 import { applyOptikToScene } from './ui/optik';
 import * as groundObjects from './ground-objects';
 import { handleLobbyMessage } from './ws-lobby-handlers';
+
+import { setMessageHandler } from './ws-connection-client';
+import { applyUndoStateChange } from './ws-undo-state';
+
 /** Mappt eine runtime-Figure auf das serialisierbare ExportFigure-Format. */
 function _toExportFig(fig: any): ExportFigure {
   return {
@@ -22,27 +26,7 @@ function _toExportFig(fig: any): ExportFigure {
     ownerId: fig.ownerId,
   };
 }
-// ── T000470: Undo/Redo-Stack-Status ─────────────────────────────────────
-export const undoState = {
-  canUndo: false,
-  canRedo: false,
-  undoCount: 0,
-  redoCount: 0,
-};
-let onUndoStateChange: ((state: typeof undoState) => void) | null = null;
-export function setUndoStateChangeHandler(fn: typeof onUndoStateChange): void {
-  onUndoStateChange = fn;
-}
 
-function applyUndoStateChange(
-  canUndo: boolean, canRedo: boolean, undoCount: number, redoCount: number,
-): void {
-  undoState.canUndo = canUndo;
-  undoState.canRedo = canRedo;
-  undoState.undoCount = undoCount;
-  undoState.redoCount = redoCount;
-  if (onUndoStateChange) onUndoStateChange({ ...undoState });
-}
 // ── Lobby/presence/session state (pure reducer) ─────────────────────────────
 let lobbyState: LobbyState = createLobbyState();
 export function getLobbyState(): LobbyState { return lobbyState; }
@@ -90,100 +74,7 @@ export function setLateJoinHandler(cb: ((name: string) => void) | null): void {
   lateJoinHandler = cb;
 }
 
-function send(msg: ClientMessage): void {
-  const ws = getWs();
-  if (ws && ws.readyState === WebSocket.OPEN) {
-    ws.send(JSON.stringify(msg));
-  }
-}
-/** Public send for lobby/admin protocol messages. */
-export function sendClient(msg: ClientMessage): void {
-  send(msg);
-}
-/** True iff a socket exists and is OPEN (used to decide sync-send vs open-hook). */
-export function isWsOpen(): boolean {
-  const ws = getWs();
-  return !!ws && ws.readyState === WebSocket.OPEN;
-}
 
-export function sendMove(id: string, x: number, z: number, facingY: number): void {
-  send({ type: 'move', id, x, z, facingY });
-}
-export function sendJump(id: string): void {
-  send({ type: 'jump', id });
-}
-export function sendUpdate(fig: any, changes: any): void {
-  send({ type: 'update', id: fig.id, changes });
-}
-export function sendStiffness(value: number): void {
-  send({ type: 'stiffness', value });
-}
-export function sendDelete(): void {
-  if (STATE.selectedId) {
-    send({ type: 'delete', id: STATE.selectedId });
-  }
-}
-export function sendUndo(): void {
-  send({ type: 'session_undo' });
-}
-export function sendRedo(): void {
-  send({ type: 'session_redo' });
-}
-export function sendAddFigure(fig: any): void {
-  send({
-    type: 'add',
-    figure: {
-      id: fig.id,
-      x: fig.root.position.x,
-      z: fig.root.position.z,
-      facingY: fig.facingY,
-      label: fig.label,
-      color: fig.color,
-      appearance: fig.appearance
-    }
-  });
-}
-
-// One-shot callback after WS OPEN + `join` frame (FE-1/REG-4 bootstrap)
-let onWsOpen: (() => void) | null = null;
-export function setWsOpenHandler(fn: (() => void) | null): void {
-  onWsOpen = fn;
-}
-export function connectWS(): void {
-  // REG-2: idempotent — never open a second socket if one is already
-  // CONNECTING/OPEN. The lobby bootstrap (main.ts) opens the socket as soon as the
-  // room is known, and bootBoard() later also calls connectWS() when the board
-  // mounts; without this guard that would create a duplicate connection.
-  const existing = getWs();
-  if (existing && (existing.readyState === WebSocket.CONNECTING || existing.readyState === WebSocket.OPEN)) {
-    return;
-  }
-  // Thread room + (when known) the canonical identity into the /sync handshake so
-  // the late-join guard (shouldRejectReconnect) can distinguish a true reconnect
-  // of an already-active player from a genuine late-joiner. Omit playerId when
-  // unknown/anon (server treats null as "not previously in room" → admit).
-  const roomFromUrl = new URLSearchParams(location.search).get('room') || 'default';
-  const params = new URLSearchParams({ room: roomFromUrl });
-  if (currentUser.userId && currentUser.userId !== 'anon') {
-    params.set('playerId', currentUser.userId);
-  }
-  const ws = new WebSocket(`${location.protocol === 'https:' ? 'wss:' : 'ws:'}//${location.host}/sync?${params.toString()}`);
-  setWs(ws);
-  (window as any).__brettWS = ws;
-  ws.addEventListener('open', () => {
-    setWsReady(true);
-    send({ type: 'join', room: roomFromUrl });
-    // FE-1/REG-4: fire the bootstrap's one-shot hook (e.g. admin_session_create)
-    // only AFTER the socket is OPEN and `join` is sent, so it never races the
-    // handshake.
-    if (onWsOpen) onWsOpen();
-  });
-  ws.addEventListener('close', () => {
-    setWsReady(false);
-    setTimeout(connectWS, 2000);
-  });
-  ws.addEventListener('message', onWsMessage);
-}
 // Injected to avoid cycle with appearance.ts.
 let applyAppearanceToFig: (fig: any, a: any) => void = () => {};
 export function setApplyAppearance(fn: typeof applyAppearanceToFig): void {
@@ -565,3 +456,13 @@ export function onWsMessage(evt: MessageEvent): void {
       break;
   }
 }
+
+setMessageHandler(onWsMessage);
+
+// Re-exports für Rückwärtskompatibilität
+export { undoState, setUndoStateChangeHandler } from './ws-undo-state';
+export {
+  sendClient, isWsOpen, sendMove, sendJump, sendUpdate, sendStiffness,
+  sendDelete, sendUndo, sendRedo, sendAddFigure, setWsOpenHandler, connectWS,
+} from './ws-connection-client';
+
